TVG-102 Blueprint for User endpoints
