+++ api_blueprints/__init__.py
@@ -4,3 +4,4 @@ from api_blueprints.search_items import search_items_blueprint
 from api_blueprints.search_subscriptions import search_subscription_blueprint
 from api_blueprints.shows import shows_blueprint
 from api_blueprints.show_episodes import show_episodes_blueprint
+from api_blueprints.user import user_blueprint
