+++ api_blueprints/user.py
@@ -1,0 +1,65 @@
+from flask import Blueprint, request
+from flask_cors import CORS
+from flask_jwt_extended import get_current_user, jwt_required
+from sqlalchemy.orm import Session
+
+from database import engine
+from database.models import User
+
+user_blueprint = Blueprint("user_blueprint", __name__)
+
+CORS(user_blueprint)
+
+@user_blueprint.route("", methods=['GET'])
+def get_user(username: str):
+    session = Session(engine)
+    user = User.search_for_user(username, session)
+    if user:
+        return user.to_dict()
+    return {'message': f'An account with the username {username} could not be found'}, 404
+
+@user_blueprint.route("/change_password", methods=['PUT'])
+@jwt_required()
+def change_password(username: str):
+    current_user: User = get_current_user()
+    session = Session(engine)
+    user = User.search_for_user(username, session)
+    if user and current_user.username == user.username:
+        user.change_password(request.json['password'])
+        session.commit()
+        return user.to_dict()
+    return { 'message': "You are not authorised to change this user's password" }, 403
+
+@user_blueprint.route("/promote", methods=['PATCH'])
+@jwt_required()
+def promote_user(username: str):
+    session = Session(engine, expire_on_commit=False)
+    current_user: User = get_current_user()
+    if current_user.role == 'Admin':
+        user = User.search_for_user(username, session)
+        if user:
+            user.promote_role()
+            session.commit()
+            session.close()
+            return '', 204
+        session.close()
+        return { 'message': f"Unable to find the user '{username}'" }, 404
+    return { 'message': 'You are not authorised to promote this user to an admin role' }, 403
+
+@user_blueprint.route("", methods=['DELETE'])
+@jwt_required()
+def delete_user(username: str):
+    session = Session(engine)
+    current_user: User = get_current_user()
+    user = User.search_for_user(username, session)
+    if current_user.username == username:
+        user.delete_user(session)
+        return { 'message': 'Your account has been deleted' }
+    elif current_user.role == 'Admin':
+        if user:
+            user.delete_user(session)
+            return { 'message': 'The account has been deleted' }
+        else:
+            return {'message': f"An account with the username '{username}' could not be found"}, 404
+    else:
+        return { 'message': 'You are not authorised to delete this user account' }
